ssupport/requirements.txt
discord.py>=2.0.0
flask>=2.0.0
requests>=2.27.0
python-dotenv>=1.0.0